Add files via upload
The code is a simple counter of word, line and characters into a file .txt but there are some features that need to enhancement. There are files for test the code in this repository. 
This command is similar to wc command in UNIX that count word, line and characters in a file. But, my code can be improved in some features.
Special characters is not consider.
And when there is a float data the program count the character like individual char. For example, 
If there are data like this 2.1 the code count 2, . and  1 in total 3 characters. I mean, it does not count this data together.